remove papercolor-theme copyright
This project actually doesn't share any code with the papercolor-theme
project that inspired this. The extent of copying now goes no further
than a couple of 256bit colour codes.

The copyright notice was originally there because this project intended
to be a clone of papercolor. This is no longer the case.